LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/window-open-noopener-expected.txt

CONSOLE MESSAGE: testError: special target must be one of: _self, _parent, _top

Harness Error (FAIL), message = testError: special target must be one of: _self, _parent, _top
